Rework Verilog module IO to produce syntactically correct Verilog
See [0] about FORMAT's directives. There a lot of them and they are
complicated. The ~{...~} allows for us to do things like Scheme's
string-join.

Mark that generated HDL code outputs for testing should also be tested
against a simulator that can perform linting. Using a linting step in
the output tests will ensure that we always produce SYNTACTICALLY valid
HDL, even if the generated output is semantically nonsensical.

Verilator has a large number of linting checks that it performs, which
makes it ideal for Verilog.

[0] https://en.wikipedia.org/wiki/Format_(Common_Lisp)#Directives