Add a "Known Problems" part to building section of README
There are some things that SBCL prints out when compiling everything
that I had to discover by hand. In particular, check-it uses a
deprecated function from asdf3, which should not be used anymore. I had
to go investigating to figure out if it was my asdf definition or
a dependent package that had it.