Add slots to every defclass so :documentation works
According to the CL specification, there must ALWAYS be a set of
parentheses for slots, even if there are no slots defined for the class.

Syntax From LispWorks HyperSpec [1]:
(defclass class-name ({superclass-name}*) ({slot-specifier}*) [[class-option]])

class-option ::= (:default-initargs . initarg-list) |
                 (:documentation string) |
                 (:metaclass class-name)

Note how the () are included on both superclass-name and
slot-specifier. Those parens MUST be included. Meanwhile, the class
options are enclosed in the optional brackets. If you do not include the
slots, then parsing these class-options are messed up and treated as
slots, which just does not work.

[1] https://www.lispworks.com/documentation/HyperSpec/Body/m_defcla.htm