Turn off Verilator's DECLFILENAME for property-based tests
Since we will be randomly generating Verilog modules and are using the
with-temporary-file macro to handle creating/removing/setting-up the
file for output & linting (which does not support providing a name), we
just disable the checking of a module's main file having a matching file
name.

This lint check is trivial to construct with smaller manual tests
anyways, so we are not terribly concerned about this.