Create class for verilog module parameters
This also means that we do not need to keep the "invalid parameters"
test. If you attempt to build an instance of the parameter class with an
invalid type for the value slot, you will receive a type error (if your
safety and/or debug is set to a high enough level).