chil/tests asdf system always uses highest safety & debug settings
I saw different behavior between my Sly instance of SBCL and when
running SBCL through make. The make command uses --no-userinit, which
tells SBCL to not read any .sbclrc file that may be present. My personal
.sbclrc has both debug and safety set to their highest level (3). This
causes type-errors to become proper errors instead of restarts (or
something, I'm not sure).

Since I want Chil to be as portable as possible, I thought it would be
best for the optimization proclamation to be made at the ASDF
level. While this does mean that tests will not be terribly fast, they
are unit & property tests and are intended to find regressions of
behavior, which requires maximum visibility into the test & code. I am
willing to pay for the slower code execution if it means figuring out
why a test fails is easier.