REBASE & SPLIT ME: Implement our interpreter for Chilog/Datalog
This is a simple lazy (hopefully) depth-first search for
substitutions. We heavily use the loop macro to keep performance up and,
particularly, to prevent exhausting the call stack, since CL does not
have tail-call optimization.

TODO: We may want to consider if putting the interpreter in its own
package is the right move. It logically makes sense, since we could
develop multiple "backends" for executing Chilog/Datalog, i.e. compiling
the Chilog to CL and re-eval-ing or shelling out to SQL. Ideally, there
is a generally abstract notion for how to handle & use Datalog that can
have different implementations "slipped under".