Allow predicate->atom to have interactive restarts for REPL work
When doing REPL work, it is convenient to be able to convert a predicate
to an atom for use on the RHS of a rule. But when going by-hand, it is
easy to screw up and have too many or too few arguments for a particular
predicate. Here we add a restart-able condition and handler to allow you
to mess up and then provide a new set of terms on-the-fly.