Rework fact and predicate database tables to be an opaque class
With this, we can present the tables abstractly. If you want to use the
tables and their values, you need to use some kind of conversion
function. This should allow us to change internal details of the
database "tables" for one reason or another while keeping a layer of
separation between us and the exact implementation of each table.