Add logging to naive interpreter
We add some low-priority logging to the naive evaluator. In theory,
this should help us track how the evaluation of a Datalog IDB to find
the full EDB goes. Most of this logging is at the debug or trace level,
since users will not want to see this much output.