Implement Chilog language
This implementation is ported from the Python in [1] to Common Lisp.

Use assert in predicate->atom to have interactive restarts REPL work
When doing REPL work, it is convenient to be able to convert a predicate
to an atom for use on the RHS of a rule. But when going by-hand, it is
easy to screw up and have too many or too few arguments for a particular
predicate. We use assert in predicate->atom to have interactive restarts
work in the REPL, allowing you to mess up and then provide a new set of
terms from the REPL. Similarly, you can do this for adding facts to a
rule.

Use adjoin instead of cons for set operations. adjoin has the benefit of
not inserting to the set when the set already contains the element to
insert.  We are allowed to TRY to add a duplicate new fact to the
provided predicate, but adjoin uses the provided :test function to
determine if elements are actually the same and if the new element
should be inserted.  The invariant we want to ACTUALLY enforce is that
we do not ADD duplicate information. This prevents us from adding
identical facts for instance.

An important feature of Chilog is the ability to generate unique
placeholder variables using a function. This function behaves like
gensym, and does NOT intern anything. YOU must capture the placeholder
value if you want to hold onto it.

We validate that a Chilog predicate's constructor arguments are
valid. For example, if a predicate is foo(X, Y), then you cannot define
a fact where you are missing information, because you would be
specifying a rule otherwise. Below are all examples of facts that we
expect to FAIL:
  foo(X, Y) :- .     -- No literals, no RHS, should be rule.
  foo(X, "baz") :- . -- Incomplete fact, what is X?
  foo("bar", Y) :- . -- Incomplete fact, what is Y?

Assert that new facts have the same length as their predicate's arity.
We also test that a built predicate uses all of the args in its head.

Also document why we use eval in Chilog type & class/object tests.

[1] https://btmc.substack.com/p/implementing-logic-programming