Implement a naive interpreter-based evaluator/engine for Chilog/Datalog
We place the interpreter in its own package because I plan to support
multiple Chilog/Datalog evaluators. For example, a semi-naive evaluator,
compiling the Chilog to CL and re-eval-ing, or shelling out to
SQL. Ideally, there is a generally abstract notion for how to handle &
use Datalog that can have different implementations "slipped under".

This first evaluator is a simple and naive depth-first search for
substitutions. In Datalog, this is a naive evaluator, because we
re-evaluate all found facts for all relations on every iteration of the
fixed-point/unification.

This naive evaluator heavily uses the loop macro to keep performance up
and, more importantly, to prevent exhausting the call stack, since CL
does not have tail-call optimization. Some of these loop calls make use
of more advanced features of LOOP, like destructuring at iteration
variable-assignment time to "pattern match" on lists of pairs, and
parallel list iteration (instead of zipping then destructuring).

Define utility functions to convert an alist to a substitution

One curious thing about this implementation is that the collect in
search-db's LOOP gets flattened. Flatten the resulting list from
searching the DB. We need to do this for 2 primary reasons:

  1. search-db returns lists and when search-db is called recursively,
     collect will collect the list of substitutions into another list
     layer. So the amount of recursion in a rule's search in the DB will
     produce a nested list-of-list-of-...-list-of hash-tables. Flatten
     flattens this out to a single list.
  2. By virtue of flatten-ing, any 'nil that occurs because a
     predicate's facts set is empty will be erased. This prevents DB
     searching when a rule's fact list is empty from returning a list of
     'nil substitutions.