Add TODO for marking high-level hardware components as debug
This idea stems from BOOM using `val debug = Output(...)` in many
modules so that bundles are not optimized out and simulators can produce
waveforms for those elements so they may be introspected. However, the
do complicate reading the module itself, since they are mostly
passed-through. I think it would be interesting to state that some
pieces are marked as debug-only so that you can choose whether or not to
hook them up when instantiating the module.